Test Suites dashboard should support biweekly and monthly ranges (#73)
- Add "biweekly" and "monthly" options to the Test Runs and Tests page
- Update the Test Runs chart to better support more than 7 days worth of data
- Update GraphQL queries to fetch workspace retention limits
- Both Test Runs and Tests pages respect retention limits (by not linking to recordings that are too old to be viewed safely)
- Tests page no longer displays past week or executions, but instead will only fetch the past hour or the past day (if chosen)
- Dropdown labels changed slightly for readability (e.g. "Last 7 days" -> "Past week")
- Add new e2e tests that cover retention behavior

---

The screenshot below shows what the test runs chart looks like with 30 days worth of data _and_ the message it displays when recordings are too old to be viewed safely (given the retention limit):
![Screenshot 2024-04-30 at 9 58 30 AM](https://github.com/replayio/dashboard/assets/29597/acfc8d7e-5ed2-45ba-ba5d-e90ee0169bf0)

The screenshot below shows the inline message we show on the Tests page when we don't fetch the full biweekly or monthly test executions (because the query is too slow to rely on):
![Screenshot 2024-04-30 at 12 58 12 PM](https://github.com/replayio/dashboard/assets/29597/a9f2e8f0-4b23-4b7a-afc5-309ffcdea49f)

+++ src/pageComponents/team/id/tests/RecordingRow.tsx
@@ -1,29 +1,50 @@
 import { Icon } from "@/components/Icon";
 import { TestSuiteTestExecutionRecording, TestSuiteTestStatus } from "@/graphql/types";
+import { getRelativeDate } from "@/utils/date";
 import { getURL } from "@/utils/recording";
 import { getColorClassName } from "@/utils/test-suites";
 import Link from "next/link";
 
 export function RecordingRow({
   recording,
+  retentionLimit,
   status,
 }: {
   recording: TestSuiteTestExecutionRecording;
+  retentionLimit: number;
   status: TestSuiteTestStatus;
 }) {
   const url = getURL(recording.id, recording.buildId);
 
-  const colorClassName = getColorClassName(status);
   const iconType = recording.isProcessed ? "processed-recording" : "unprocessed-recording";
 
-  return (
-    <Link
-      className="flex flex-row items-center gap-2 text-white px-2 py-1 hover:bg-slate-700"
-      data-test-name="RecordingRow"
-      href={url ?? ""}
-    >
-      <Icon className={`w-6 h-6 shrink-0 ${colorClassName}`} type={iconType} />
-      <div className="truncate shrink">View recording</div>
-    </Link>
-  );
+  const isWithinRetentionLimit =
+    recording.createdAt.getTime() >= getRelativeDate({ daysAgo: retentionLimit }).getTime();
+
+  if (isWithinRetentionLimit) {
+    const colorClassName = getColorClassName(status);
+
+    return (
+      <Link
+        className="flex flex-row items-center gap-2 text-white px-2 py-1 hover:bg-slate-700"
+        data-status-viewable
+        data-test-name="RecordingRow"
+        href={url ?? ""}
+      >
+        <Icon className={`w-6 h-6 shrink-0 ${colorClassName}`} type={iconType} />
+        <div className="truncate shrink">View recording</div>
+      </Link>
+    );
+  } else {
+    return (
+      <div
+        className="flex flex-row items-center gap-2 text-white px-2 py-1"
+        data-status-not-viewable
+        data-test-name="RecordingRow"
+      >
+        <Icon className="w-6 h-6 shrink-0 text-slate-400" type={iconType} />
+        <div className="truncate shrink text-slate-400">Recording not available</div>
+      </div>
+    );
+  }
 }
